Refactoring generation handlers
Addresses #41.  Created a new class AnnotationProvider which wraps the
annotation collection and provides better accessor methods to the collection
internals.  Handlers now are responsible for taking the context (which includes
the Body object) and applying the correct generated code.  They are wrapped in
a stack which loops over all of the handlers.

Signed-off-by: Nate Brunette <n@tebru.net>